apps/robot/requirements.txt
# KolayRobot Robot - Python Dependencies
# Linux sunucuda native çalışacak

# Web Automation
playwright==1.49.0

# Database (MySQL)
pymysql==1.1.0
aiomysql==0.2.0
cryptography==42.0.2
sqlalchemy==2.0.25
alembic==1.13.1

# Email
imapclient==3.0.1
email-validator==2.1.0

# Excel/CSV Processing
openpyxl==3.1.2
pandas==2.1.4

# Async & Concurrency
asyncio-mqtt==0.16.2
aiofiles==23.2.1
celery==5.3.6
redis==5.0.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging
loguru==0.7.2
rich==13.7.0

# Utilities
python-dateutil==2.8.2
tenacity==8.2.3

# API (for admin panel communication)
fastapi==0.109.0
uvicorn==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Scheduling
apscheduler==3.10.4
